Refactor: extract shared types, validation, and error handling
- Extract shared Graph API types (graphEmailBody, graphMailAddress,
  graphDateTimeZone, graphEventLocation, graphAttendee) into types.go
  to eliminate duplication across action files
- Extract error handling (graphError, mapGraphError) into response.go,
  matching the pattern used by the github connector
- Extract validateEmail and detectContentType into validation.go so
  they can be reused across actions (send_email and create_calendar_event)
- Add attendee email validation to create_calendar_event (was only
  validating to/cc in send_email — inconsistent)
- Add test for invalid attendee email validation
- Update README.md file structure section
