+++ connectors/microsoft/README.md
@@ -185,6 +185,9 @@ When adding a new action, add it to the `Manifest()` return value with a `Parame
 ```
 connectors/microsoft/
 ├── microsoft.go                  # MicrosoftConnector struct, New(), Manifest(), doRequest(), ValidateCredentials()
+├── types.go                      # Shared Microsoft Graph API types (graphEmailBody, graphMailAddress, etc.)
+├── response.go                   # Graph API error response → typed connector error mapping
+├── validation.go                 # Shared validation helpers (validateEmail, detectContentType)
 ├── send_email.go                 # microsoft.send_email action
 ├── list_emails.go                # microsoft.list_emails action
 ├── create_calendar_event.go      # microsoft.create_calendar_event action
