+++ connectors/microsoft/validation.go
@@ -1,0 +1,28 @@
+package microsoft
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/supersuit-tech/permission-slip-web/connectors"
+)
+
+// validateEmail performs basic email address validation for a recipient list entry.
+// Used by send_email (to/cc) and create_calendar_event (attendees).
+func validateEmail(field string, idx int, addr string) error {
+	if addr == "" {
+		return &connectors.ValidationError{Message: fmt.Sprintf("%s[%d] is empty", field, idx)}
+	}
+	if !strings.Contains(addr, "@") || strings.HasPrefix(addr, "@") || strings.HasSuffix(addr, "@") {
+		return &connectors.ValidationError{Message: fmt.Sprintf("%s[%d] is not a valid email address: %q", field, idx, addr)}
+	}
+	return nil
+}
+
+// detectContentType returns "HTML" if the body contains HTML tags, otherwise "Text".
+func detectContentType(body string) string {
+	if strings.Contains(body, "<") && strings.Contains(body, ">") {
+		return "HTML"
+	}
+	return "Text"
+}
